fix(bundle): implement generic bundle parsing
